+++ src/main/java/org/patternfly/component/TemplateComponent.java
@@ -49,6 +49,11 @@ public class TemplateComponent extends BaseComponent<HTMLElement, TemplateCompon
         return this;
     }
 
+    // override to assure internal wiring
+    public TemplateComponent add(/* Foo foo */) {
+        return this;
+    }
+
     // ------------------------------------------------------ builder
 
     public TemplateComponent methodsReturningAReferenceToItself() {