+++ src/main/java/org/patternfly/component/avatar/Avatar.java
@@ -13,7 +13,12 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.patternfly.component;
+package org.patternfly.component.avatar;
+
+import org.patternfly.component.BaseComponentFlat;
+import org.patternfly.component.ComponentType;
+import org.patternfly.layout.Brightness;
+import org.patternfly.layout.Size;
 
 import elemental2.dom.HTMLImageElement;
 
@@ -22,13 +27,11 @@ import static org.patternfly.layout.Classes.avatar;
 import static org.patternfly.layout.Classes.component;
 
 /**
- * PatternFly avatar component.
+ * An avatar is a visual used to represent a user. It may contain an image or a placeholder graphic.
  *
- * @see <a href=
- *      "https://www.patternfly.org/v4/documentation/react/components/avatar/">https://www.patternfly.org/v4/documentation/react/components/avatar</a>
+ * @see <a href= "https://www.patternfly.org/components/avatar/html">https://www.patternfly.org/components/avatar/html</a>
  */
-@Deprecated
-public class Avatar extends BaseComponent<HTMLImageElement, Avatar> {
+public class Avatar extends BaseComponentFlat<HTMLImageElement, Avatar> {
 
     // ------------------------------------------------------ factory
 
@@ -39,8 +42,30 @@ public class Avatar extends BaseComponent<HTMLImageElement, Avatar> {
     // ------------------------------------------------------ instance
 
     Avatar(String src, String alt) {
-        super(img(src).css(component(avatar)).element(), "Avatar");
+        super(img().css(component(avatar)).apply(i -> {
+            i.src = src;
+            i.alt = alt;
+        }).element(), ComponentType.Avatar);
+    }
+
+    // ------------------------------------------------------ builder
+
+    public Avatar src(String src) {
+        element().src = src;
+        return this;
+    }
+
+    public Avatar alt(String alt) {
         element().alt = alt;
+        return this;
+    }
+
+    public Avatar size(Size size) {
+        return css(size.modifier);
+    }
+
+    public Avatar border(Brightness brightness) {
+        return css(brightness.modifier);
     }
 
     @Override